+++ km3io/definitions/fitparameters.py
@@ -1,0 +1,27 @@
+# -*- coding: utf-8 -*-
+"""
+KM3NeT Data Definitions v1.1.2
+https://git.km3net.de/common/km3net-dataformat
+"""
+
+# fitparameters
+data = {
+    "JGANDALF_BETA0_RAD": 0,
+    "JGANDALF_BETA1_RAD": 1,
+    "JGANDALF_CHI2": 2,
+    "JGANDALF_NUMBER_OF_HITS": 3,
+    "JENERGY_ENERGY": 4,
+    "JENERGY_CHI2": 5,
+    "JGANDALF_LAMBDA": 6,
+    "JGANDALF_NUMBER_OF_ITERATIONS": 7,
+    "JSTART_NPE_MIP": 8,
+    "JSTART_NPE_MIP_TOTAL": 9,
+    "JSTART_LENGTH_METRES": 10,
+    "JVETO_NPE": 11,
+    "JVETO_NUMBER_OF_HITS": 12,
+    "JENERGY_MUON_RANGE_METRES": 13,
+    "JENERGY_NOISE_LIKELIHOOD": 14,
+    "JENERGY_NDF": 15,
+    "JENERGY_NUMBER_OF_HITS": 16,
+    "JCOPY_Z_M": 17,
+}