+++ km3io/definitions/trigger.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+"""
+KM3NeT Data Definitions v1.1.2
+https://git.km3net.de/common/km3net-dataformat
+"""
+
+# trigger
+data = {
+    "JTRIGGER3DSHOWER": 1,
+    "JTRIGGERMXSHOWER": 2,
+    "JTRIGGER3DMUON": 4,
+    "JTRIGGERNB": 5,
+}