+++ tests/test_offline.py
@@ -184,6 +184,16 @@ class TestOfflineReader(unittest.TestCase):
         self.assertEqual(best.size, 9)
         self.assertEqual(best['JGANDALF_BETA1_RAD'][:4].tolist(), JGANDALF_BETA1_RAD)
 
+    def test_reading_header(self):
+        # head is the supported format
+        head = OfflineReader(OFFLINE_NUMUCC).header
+
+        self.assertEqual(float(head['DAQ']), 394)
+        self.assertEqual(float(head['kcut']), 2)
+
+        # test the warning for unsupported fheader format
+        self.assertWarns(UserWarning, self.r.header,
+                         "Your file header has an unsupported format")
 
 
 class TestOfflineEvents(unittest.TestCase):